Clean up documentation.
A lot of documentaiton (like the code) is coupled together. We still need to move some documentation over... but some of that code hasn't moved yet. i.e. code for Jobs hasn't moved yet, so the documentation probably shouldn't either.

For now I'll leave the DataHub documentation untouched, so things are clearer until this is totally cleaned up.