Continously release versions.
This is done via a custom GitHub workflow / action that uses shipkit github release and auto version plugins. This means we will publish a release on GitHub and bintray for every commit on a branch named `release/*`.

For now autopublish to bintray is still disabled, we can enable it once we're sure we're happy with this pipeline.