Start publishing jars to bintray.
For now this is a manual process. We should be able to easily hook this up to a CI/CD pipeline via github workflows.

This manual process is documentedin ci-release.gradle in comments. We should be able to hook up a CI/CD pipeline very quickly. I didn't do it in this PR A) to narrow the scope and B) so we can debate on some minor jar naming and stuff before releasing, if we want to.