Continously release versions. (#12)
This is done via a custom GitHub workflow / action that uses shipkit github release and auto version plugins. This means we will publish a release on GitHub and bintray for every push of the master branch.

For now autopublish to bintray is still disabled, we can enable it once we're sure we're happy with this pipeline.