Use ElasticSearch in single-node mode (#76)
We've been seeing datahub integration tests fail on Linux machines (e.g. linkedin/datahub#2088). The core reason for this issue is that production deployments of ElasticSearch require some extra configuration on Linux machines.

However, we don't need a production deployment for integration tests, and as per the ES docs, we can use single-node mode to bypass these checks and make the tests succeed on Linux-based distros.

Note that this is already done in the es-7 directories