Add a topic convention class for MXE v5, which determines how to name MXE v5 event topics and get their schemas.
We can then use this to dynamically get topic names (and schemas) from Urn/RecordTemplates, rather than hard code everything, which we have done internally.

This ships with a default implementation that allows for some customization, but without any will do %event type%_%entity%_%aspect%.

This is being split off from the one in DataHub, which was both v4 and v5 (but v5 was not implemented there; it is here, we can delete it there in a follow up).

https://github.com/linkedin/datahub/tree/125ae288f1601bef2c42f44db6d8eb58b380a304/metadata-events/mxe-registration/src/main/java/com/linkedin/mxe