Fix resource loading in the ES test framework. (#98)
For static indices, we would end up calling testInstance.getClass().getClass().getResource() (indirectly). Somehow java 8 is fine with this and can locate the resource (off java.lang.Class), but java 11 is not okay with it (which is what I would expect).

Fix the code by being explicit about which class to use.