(feat) Add support for atomic update of multiple aspects.
At the moment, the BaseLocalDAO only supports updating a single aspect
at a time, meaning that, if multiple aspects are updated as a result of
a single HTTP request to the GMS, the update is not atomic. This commit
adds an `addMany` method to BaseLocalDAO, which offers to (configurable)
ability to update multiple aspects in a single transaction.