Add toggle-able logic for job-gms record retrieval (#252)
<feat> [getLatest()]: add support to use two alternative methods for checking record existence during record insertion

Based off of logs introduced in a previous changeset, there is good reason to believe that Ebean's EbeanServer.find(Class, Object) method does not work accurately at scale, returning no matching existing records when it is clear that such a record exists. This appears to be the cause of mass duplicity introduced into the database. Luckily, Ebean has several ways to accomplish a task such as checking record existence, so this changeset attempts to leverage two of these in attempt to stop further duplicates from being introduced.

Might introduce breaking change in future if either (or both) of the alternative pathways is removed after adoption.

---------

Co-authored-by: Jonathan Hui <jhui@jhui-ld2.linkedin.biz>